refactor: move dynamicTitle into separated file
