+++ src/chrome-scripts/features/service-worker/index.js
@@ -1,0 +1,1 @@
+export { default as dynamicTitle } from './dynamicTitle.js';
